Closed areas is now a hashtable of grid locations instead of being cells with flag set.
Cell no longer has IsEmpty(), IsEnclosed(), SetEnclosed(...), nor NothingToDraw(). If there is a cell being occupied, it is assumed not empty, and also never has nothing to draw. It is enclosed if it is in the enclosed list.

Cell no longer fills, instead taking the Surface (which will soon be changes to be a MapFixture) in the constructor.

Some performance improvement when checking for closed areas.