Added overload to DrawBox(...) to take a clipping rectangle, with assertions on the clipping rectangle.
Added overload to DrawBorderedBox(...) to take two rectangles for the outer border dimensions and the inner dimensions. This can be used to draw uneven and / or offset borders.

Added another overload to above DrawBorderedBox(...) to also taken a clipping rectangle. This new overload, also taking two rectangles (for the box and border) is also now the base logic.

Updated existing DrawSprite(...) overloads to get the screen clipping rectangle not the screen rectangle, with assertions on the clipping rectangle and more on the source rectangle. Using the screen clipping rectangle leads to the base logic being more consistant.